+++ ChangeLog.md
@@ -7,6 +7,7 @@
     Release date: 
 
     - When choosing a node for a service, make sure the current node is the first candidate
+    - Databases list must be filtered through the configured list
 
 ### 4.11.0